Refactoring, file split
