+++ packages/web-console/src/scenes/Editor/Monaco/questdb-sql/createSchemaCompletionProvider.ts
@@ -1,93 +1,12 @@
 import { Table, uniq } from "../../../../utils"
 import * as monaco from "monaco-editor"
-import { InformationSchemaColumn } from "./types"
-import { editor, IRange } from "monaco-editor"
-import { languages } from "monaco-editor"
+import { CompletionItemPriority, InformationSchemaColumn } from "./types"
+import { editor } from "monaco-editor"
 import IStandaloneCodeEditor = editor.IStandaloneCodeEditor
 import { findMatches, getQueryFromCursor } from "../utils"
-import { operators } from "./operators"
-import { dataTypes, functions, keywords } from "@questdb/sql-grammar"
-
-const getLanguageCompletions = (range: IRange) => [
-  ...functions.map((qdbFunction) => {
-    return {
-      label: qdbFunction,
-      kind: languages.CompletionItemKind.Function,
-      insertText: qdbFunction,
-      range,
-    }
-  }),
-  ...dataTypes.map((item) => {
-    return {
-      label: item,
-      kind: languages.CompletionItemKind.Keyword,
-      insertText: item,
-      range,
-    }
-  }),
-  ...keywords.map((item) => {
-    const keyword = item.toUpperCase()
-    return {
-      label: keyword,
-      kind: languages.CompletionItemKind.Keyword,
-      insertText: keyword,
-      range,
-    }
-  }),
-  ...operators.map((item) => {
-    const operator = item.toUpperCase()
-    return {
-      label: operator,
-      kind: languages.CompletionItemKind.Operator,
-      insertText: operator.toUpperCase(),
-      range,
-    }
-  }),
-]
-
-export const getColumnCompletions = (
-  columns: InformationSchemaColumn[],
-  range: IRange,
-  withTableName?: boolean,
-) => {
-  // For JOIN ON ... completions, return `table.column` text
-  if (withTableName) {
-    return columns.map((item) => ({
-      label: {
-        label: `${item.table_name}.${item.column_name}`,
-        detail: "",
-        description: item.data_type,
-      },
-      kind: languages.CompletionItemKind.Enum,
-      insertText: `${item.table_name}.${item.column_name}`,
-      sortText: "1",
-      range,
-    }))
-    // For everything else, return a list of unique column names.
-  } else {
-    return uniq(columns.map((item) => item.column_name)).map((columnName) => {
-      const tableNames = columns
-        .filter((item) => item.column_name === columnName)
-        .map((item) => item.table_name)
-      return {
-        label: {
-          label: columnName,
-          detail: ` (${tableNames.sort().join(", ")})`,
-          // If the column is present in multiple tables, show their list here, otherwise return the column type.
-          description:
-            tableNames.length > 1
-              ? ""
-              : columns.find((item) => item.column_name === columnName)
-                  ?.data_type,
-        },
-        kind: languages.CompletionItemKind.Enum,
-        insertText: columnName,
-        sortText: "1",
-        range,
-      }
-    })
-  }
-}
+import { getTableCompletions } from "./getTableCompletions"
+import { getColumnCompletions } from "./getColumnCompletions"
+import { getLanguageCompletions } from "./getLanguageCompletions"
 
 export const createSchemaCompletionProvider = (
   editor: IStandaloneCodeEditor,
@@ -149,20 +68,6 @@ export const createSchemaCompletionProvider = (
           const openQuote = textUntilPosition.substr(-1) === '"'
           const nextCharQuote = nextChar == '"'
 
-          const tableSuggestions = tables.map((item) => {
-            return {
-              label: item.table_name,
-              kind: languages.CompletionItemKind.Class,
-              insertText: openQuote
-                ? item.table_name + (nextCharQuote ? "" : '"')
-                : /^[a-z0-9_]+$/i.test(item.table_name)
-                ? item.table_name
-                : `"${item.table_name}"`,
-              sortText: "1",
-              range,
-            }
-          })
-
           if (
             /(FROM|INTO|(ALTER|BACKUP|DROP|REINDEX|RENAME|TRUNCATE|VACUUM) TABLE|JOIN|UPDATE)\s$/gim.test(
               textUntilPosition,
@@ -171,7 +76,13 @@ export const createSchemaCompletionProvider = (
               !textUntilPosition.endsWith("= '"))
           ) {
             return {
-              suggestions: tableSuggestions,
+              suggestions: getTableCompletions({
+                tables,
+                range,
+                priority: CompletionItemPriority.High,
+                openQuote,
+                nextCharQuote,
+              }),
             }
           }
 
@@ -186,21 +97,33 @@ export const createSchemaCompletionProvider = (
                 textUntilPosition.match(/\sON\s/gim) !== null
               return {
                 suggestions: [
-                  ...getColumnCompletions(
-                    informationSchemaColumns.filter((item) =>
+                  ...getColumnCompletions({
+                    columns: informationSchemaColumns.filter((item) =>
                       tableContext.includes(item.table_name),
                     ),
                     range,
                     withTableName,
-                  ),
+                    priority: CompletionItemPriority.High,
+                  }),
                   ...getLanguageCompletions(range),
                 ],
               }
             } else {
               return {
                 suggestions: [
-                  ...getColumnCompletions(informationSchemaColumns, range),
-                  ...tableSuggestions,
+                  ...getColumnCompletions({
+                    columns: informationSchemaColumns,
+                    range,
+                    withTableName: false,
+                    priority: CompletionItemPriority.High,
+                  }),
+                  ...getTableCompletions({
+                    tables,
+                    range,
+                    priority: CompletionItemPriority.MediumHigh,
+                    openQuote,
+                    nextCharQuote,
+                  }),
                   ...getLanguageCompletions(range),
                 ],
               }
@@ -210,7 +133,13 @@ export const createSchemaCompletionProvider = (
           if (word.word) {
             return {
               suggestions: [
-                ...tableSuggestions,
+                ...getTableCompletions({
+                  tables,
+                  range,
+                  priority: CompletionItemPriority.High,
+                  openQuote,
+                  nextCharQuote,
+                }),
                 ...getLanguageCompletions(range),
               ],
             }
