+++ packages/web-console/src/scenes/Editor/Monaco/questdb-sql/types.ts
@@ -4,3 +4,11 @@ export type InformationSchemaColumn = {
   column_name: string
   data_type: string
 }
+
+export enum CompletionItemPriority {
+  High = "1",
+  MediumHigh = "2",
+  Medium = "3",
+  MediumLow = "4",
+  Low = "5",
+}
